+++ NEWS.md
@@ -1,3 +1,12 @@
+# shinyGizmo 0.2.2
+
+* Add `animateVisibility` JS call to show and hide elements with animation.
+* Add `runAnimation` helper callback to run element animations.
+
+# shinyGizmo 0.2.1
+
+* Add `once` argument to `conditionalJS` that makes jsCall run only when condition value changed.
+
 # shinyGizmo 0.2
 
 * Add `conditionalJS` component - extension of `shiny::conditionalPanel` that allows to run custom JS when condition is met.